+++ src/interfaces/selectCssClasses/selectCssClasses.interface.ts
@@ -3,4 +3,10 @@
  */
 export interface SelectCssClasses
 {
+    //######################### properties #########################
+
+    /**
+     * Css class applied to container storing visual parts of Select
+     */
+    visualContainer: string;
 }